+++ CHANGELOG.md
@@ -3,7 +3,12 @@
 
 ### Features
 
-* import tailwind, vitest and components from previous next.js version ([1b45ac0](https://github.com/seanWLawrence/cobra-events/commit/1b45ac0bebf2a1e9d2de304929fc01d37e1c70af))
+* add error and catch boundaries at root level ([e39de7a](https://github.com/seanWLawrence/cobra-events/commit/e39de7a8cf0c0cecf7330fb0c9debb5d47581527))
+* import tailwind, vitest and components from previous next.js version ([830bb80](https://github.com/seanWLawrence/cobra-events/commit/830bb8004c6b1469ae7dc8158cc01421c7ffd8b6))
+
+
+
+